Add migration and managers for models
